requirements.txt
# Strands Agents
strands-agents
strands-agents-tools

# Core AWS SDK and services
boto3>=1.35.0
botocore>=1.35.0

# AWS CDK for infrastructure as code
aws-cdk-lib>=2.150.0
constructs>=10.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
ruff>=0.1.0
isort>=5.12.0
pre-commit>=3.5.0

# Security
detect-secrets>=1.4.0

# Environment management
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.2.0
